toke.c: Avoid a conversion to/from UTF-8
If the source file is encoded as UTF-8, we don't have to find its code
point equivalent when parsing--we can just copy it unchanged.  This
wasn't done before because of the fear the input would be malformed, and
finding the code point had the side effect of checking for
well-formedness.  The previous commit added wellformedness checking,
so doing it again here would be redundant.