add a load_module routine to Module-Load
The main objection to Module::Load appears to be that its load()
routine guesses at whether the user is asking to load a module or
a file.  This commit adds an unambiguous load_module() routine that
will always load the given name as a module.

Although I have copied and pasted some tests and updated them to
test load_module, I believe more testing would be useful here.

Further, I have not updated the (strange-seeming) importer to allow
the user to import load_module.